f commit

+++ controllers/ModController.js
@@ -4,6 +4,9 @@ var fs = require('fs');
 
 let makePublic = async(req,res) =>{
     try{
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         const result = await Dbquery.makePublic(req.body.questionsId);
         return res.status(200).json(result);
     } catch(error){
@@ -14,6 +17,9 @@ let makePublic = async(req,res) =>{
 
 let deleteQuestion = async(req,res) =>{
     try{
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         const result = await Dbquery.deleteQuestion(req.body.questionsId);
         return res.status(200).json(result);
     } catch(error){
@@ -22,6 +28,9 @@ let deleteQuestion = async(req,res) =>{
 }
 let getUnlive = async(req,res) =>{
     try{
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         const result = await Dbquery.getUnlive(1 * req.body.skip,req.body.limit);
         return res.status(200).json(result);
     } catch(error){
@@ -31,6 +40,9 @@ let getUnlive = async(req,res) =>{
 }
 let addCategory = async(req,res)=>{
     try{
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         const result = await Dbquery.addCategory(req.body.questionsId,req.body.category);
         return res.status(200).json(result);
     } catch(error){
@@ -39,6 +51,9 @@ let addCategory = async(req,res)=>{
 }
 const getAuto = (req,res)=>{
     try {
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         let auto = require('../config.json');   
         res.status(200).json(auto)    
     } catch (error) {
@@ -47,6 +62,9 @@ const getAuto = (req,res)=>{
 }
 const setAuto = (req,res)=>{
     try {
+        if(req.decoded['role']!='admin'){
+            return res.status(500).json('e');
+        }
         let auto = require('../config.json');  
         auto['status'] = !auto['status'];
         fs.writeFile('../config.json', JSON.stringify(auto), 'utf8',()=>{
